update android sdk version + fix html export + small fixes

+++ app/src/main/java/externius/rdmg/helpers/Utils.java
@@ -113,14 +113,14 @@ public final class Utils {
         if (monsterType.equalsIgnoreCase("any")) {
             for (Monster monster : monsters) {
                 if (parse(monster.getChallengeRating()) <= partyLevel + 2
-                        && parse(monster.getChallengeRating()) >= Math.floor((double)(partyLevel / 4))) {
+                        && parse(monster.getChallengeRating()) >= Math.floor((float)(partyLevel / 4))) {
                     result.add(monster);
                 }
             }
         } else {
             for (Monster monster : monsters) {
                 if (parse(monster.getChallengeRating()) <= partyLevel + 2
-                        && parse(monster.getChallengeRating()) >= Math.floor((double)(partyLevel / 4))
+                        && parse(monster.getChallengeRating()) >= Math.floor((float)(partyLevel / 4))
                         && monsterType.contains(monster.getType())) {
                     result.add(monster);
                 }
